+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- [ISSUE-101](https://github.com/microsoft/scitt-ccf-ledger/issues/101): Replaced workaround code with `t_cose` for signature verification of Notary style claims.
+
 ### Dependencies
 - [ISSUE-100](https://github.com/microsoft/scitt-ccf-ledger/issues/100): Update CCF from [3.0.2 to 3.0.6](https://github.com/microsoft/CCF/compare/ccf-3.0.2...ccf-3.0.6) which means a `t_cose` upgrade from [v1.1 to v1.1.1](https://github.com/laurencelundblade/t_cose/compare/v1.1...v1.1.1).